email-agent/pyproject.toml
[project]
name = "email-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.6.0",
    "ipykernel>=6.29.5",
    "langchain-core>=0.3.66",
    "litellm==1.73.6",
    "numpydoc>=1.9.0",
    "openpipe-art==0.3.12",
    "panza>=0.1.0",
    "pip>=25.1.1",
    "pydantic>=2.11.7",
    "rich>=13.9.4",
    "skypilot[runpod]>=0.9.3",
    "vllm>=0.9.1 ; sys_platform == 'linux'",
    "weave>=0.51.54",
]

[tool.uv]
override-dependencies = ["vllm; sys_platform == 'linux'"]
